+++ include/compiler_hooks.h
@@ -105,8 +105,8 @@
 
 #define def_set_tile_color               (def_player_unhook + 3)
 #define def_set_tile_pattern             (def_set_tile_color + 3)
-#define def_set_sprite_transpose         (def_set_tile_pattern + 3)
-#define def_set_sprite_color             (def_set_sprite_transpose + 3)
+#define def_set_sprite_flip         (def_set_tile_pattern + 3)
+#define def_set_sprite_color             (def_set_sprite_flip + 3)
 #define def_set_sprite_pattern           (def_set_sprite_color + 3)
 
 #define def_usr0	                     (def_set_sprite_pattern + 3)